-	Adds the ability to select between multiple windows, where each of the windows has the same executable name. This is done by adding a new, optional parameter to the win_selector rule, a “title fragment”, which tries to match the title of one of the various Windows. -	Changes the loop which looks at each window object, and filters out any Windows without either a title or executable name, along with some other heuristics based on common Windows which the user would presumably never actually want to focus (currently all Windows Explorer Windows) -	Removes the focus mimics rule entirely -- after re-factoring the window–choosing algorithm, this was no longer necessary.
Going back to the previous check-in, the original problem here had to do with the code that was selecting which window to bring to the foreground based on the focus command. Anytime that code successfully found a window (based on a focus command), it would persist the window object in memory, for quicker lookup in the future. The sort of made sense, because get_all_windows actually returns a huge number of Windows objects, most of which are hidden, and iterating through them every time you want to look up a visible window is slow, though not so slow as to be really noticeable at the UI level; that is, the focus command seems to run plenty fast even if you have to iterate through all of the hidden Windows objects to find the one you’re looking for.  Note that, it might be better if get_all_windows could filter the results to return only visible windows (instead of doing it in the python loop) – this could be an optimization.

Anyhow, I was running into a bug in which a previously persisted (in memory) window object handle had become invalid, so that the optimized, quick look up would error out when the app tried to set it to the foreground window. This seemed to happen regularly with IntelliJ idea. I best guess as to what was happening there is that switching tabs in Idea actually changed the “main visible window object” for Idea; that is, I’m guessing that each application has many different Windows objects, each of which may or may not be visible, and that an application can change the visible window that sort of “represents” the application.

I essentially fixed the main bug in the previous check-in by disabling the code that was remembering Windows objects, so that every focus command would have to iterate through all found Windows objects each time. This worked, and it doesn’t seem too slow at all.  I’m sure there’s some clever way to try to manage all of the various Windows objects for a given application, so that you could get the right one without having to iterate all the windows every time, but it certainly doesn’t seem worth the effort at this point.

Another thing to note here is about the WinSelectorRule. I’m still confused about exactly how this rule works, but I’m pretty convinced that it’s important -- I think that somehow by using the win_names_ref, the rule is actually building up a command list based on the crazy dictionary specification.