This commit comment is for this commit and the previous commit.
First of all, the languages directory no longer gets searched by natlinkmain; subsequently, the rules in that directory no longer get directly loaded by natlinkmain. The language rules have no grammar defined in the file; they get imported by _switcher, and added to the grammar for that rule/module. This is to ensure that the language rules only get loaded one time.

Each language rule also imports most of its command specifications from specs.py. So, we have an import chain like:

_switcher <- python_rule <- specs

Because the language rules and specs are not in the Macrosystem directory, they won’t get reloaded by a microphone toggle. So, there’s a special rule to reload everything during a runtime; this is defined in _reloader. The “reload languages” rule there uses a Python extension, reimport, to force the reload of specs, and then every language rule. After that, the rule “touches” _switcher, and then toggles the microphone. This ensures that everything gets reimported into _switcher, and that then NatLink reloads _switcher.

The important thing to remember in all of this is that importing a Python module is entirely different (somehow) from NatLink actually loading a rule/grammar. What can happen, if you’re not careful, is that you will end up with two runtime instances of the same rule/grammar. You can make this happen by importing a module (or a rule or grammar from that module), a module that also gets loaded by NatLink from the Macrosystem directory. Note that the behavior of something like this is unpredictable in terms of how it manifests: it depends on load order which is unordered and can change from one run to the next. So, if you’re trying to import a rule, make sure it’s not loaded by a grammar in the same module, and make sure it’s not in Macrosystem.